+++ src/client/java/nl/enjarai/doabarrelroll/fabric/DoABarrelRollFabricClient.java
@@ -1,5 +1,6 @@
 package nl.enjarai.doabarrelroll.fabric;
 
+import com.bawnorton.mixinsquared.api.MixinCanceller;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents;
 import net.fabricmc.fabric.api.client.keybinding.v1.KeyBindingHelper;
@@ -12,10 +13,13 @@ import nl.enjarai.doabarrelroll.DoABarrelRollClient;
 import nl.enjarai.doabarrelroll.ModKeybindings;
 import nl.enjarai.doabarrelroll.config.ModConfig;
 import nl.enjarai.doabarrelroll.fabric.net.HandshakeClientFabric;
+import nl.enjarai.doabarrelroll.util.MixinCancelChecker;
 import nl.enjarai.doabarrelroll.util.StarFoxUtil;
 import org.slf4j.Logger;
 
-public class DoABarrelRollFabricClient implements ClientModInitializer {
+import java.util.List;
+
+public class DoABarrelRollFabricClient implements ClientModInitializer, MixinCanceller {
     @Override
     public void onInitializeClient() {
         DoABarrelRollClient.init();
@@ -34,5 +38,15 @@ public class DoABarrelRollFabricClient implements ClientModInitializer {
 
             StarFoxUtil.clientTick(client);
         });
+    }
+
+    @Override
+    public boolean shouldCancel(List<String> targetClassNames, String mixinClassName) {
+        if (mixinClassName.equals("com.anthonyhilyard.equipmentcompare.mixin.KeyMappingMixin") && !MixinCancelChecker.hasChangedPriority(mixinClassName, 0)) {
+            DoABarrelRoll.LOGGER.warn("Equipment Compare detected, disabling their overly invasive keybinding mixin. Report any relevant issues to them.");
+            DoABarrelRoll.LOGGER.warn("If the author of Equipment Compare is reading this: see #31 on your github. Once you've fixed the issue, you can set the priority of this mixin explicitly to stop it being disabled.");
+            return true;
+        }
+        return false;
     }
 }